add metadata for home, about and frequently asked questions
